Move Hotjar event code to another file
This shortens header.js file even more
